% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_analysis_visualization.R
\name{barplot_proximity}
\alias{barplot_proximity}
\title{bar plot of spatial proximity for pairwise cell types}
\usage{
barplot_proximity(cell_proximity_df, font.size = 20)
}
\arguments{
\item{cell_proximity_df}{a dataframe stored in the output list from function "cell_proximity_enrichment_score_single" or "cell_proximity_enrichment_score_multiple", e.g., cell_proximity$CPScore_df}

\item{font.size}{font.size for ggplot}
}
\value{
a ggplot object
}
\description{
bar plot of spatial proximity for pairwise cell types
}
