% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NeuronChat_source_file.R
\name{cal_expr_by_group}
\alias{cal_expr_by_group}
\title{Calculate the average gene expression by cell group}
\usage{
cal_expr_by_group(df, gene_used, mean_method = NULL)
}
\arguments{
\item{df}{a data frame or data.table containing signaling gene expression (row: cell; column: genes; the last column is cell group labels)}

\item{gene_used}{gene symbols used to calculate average expression for cell groups}
}
\value{

}
\description{
Calculate the average gene expression by cell group
}
