% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_analysis_visualization.R
\name{cell_proximity_enrichment_score_single}
\alias{cell_proximity_enrichment_score_single}
\title{Calculation of cell proximity enrichment score for single tissue slice}
\usage{
cell_proximity_enrichment_score_single(
  meta,
  celltype_label = "subclass",
  centroid_x = "centroid_x",
  centroid_y = "centroid_y",
  thresh_dist = 400,
  permutation_number = 1000
)
}
\arguments{
\item{meta}{A dataframe containing spatial locations and cell type annotations of cells}

\item{celltype_label}{variable name (character) in meta, to indicate the cell type annotations of cells}

\item{centroid_x}{variable name (character) in meta, to indicate the x coordinates of cells}

\item{centroid_y}{variable name (character) in meta, to indicate the y coordinates of cells}

\item{thresh_dist}{distance threshold (a scalar) used to determine the spatial neighbors of cells}

\item{permutation_number}{number of permutations (a scalar) used to calculate expected frequency matrices and p-values}
}
\value{
A list contains orginal frequency matrix, permutation frequency matrices, dataframe containing cell proximity enrichment score & pvalue, CPscore matrix, pvalue (adjusted) matrix
}
\description{
Calculation of cell proximity enrichment score & associated p-values for single tissue slice
}
