% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_analysis.R
\name{computeNetSimilarity_Neuron}
\alias{computeNetSimilarity_Neuron}
\title{Compute signaling network similarity for any pair of signaling networks}
\usage{
computeNetSimilarity_Neuron(
  object,
  slot.name = "net",
  type = c("functional", "structural"),
  k = NULL,
  thresh = NULL
)
}
\arguments{
\item{object}{NeuronChat object}

\item{slot.name}{the slot name of object that is used to store communication strength matrices, i.e.,  'net'}

\item{type}{"functional"}

\item{k}{the number of nearest neighbors}

\item{thresh}{the fraction (0 to 0.25) of interactions to be trimmed before computing network similarity}
}
\value{

}
\description{
Compute signaling network similarity for any pair of signaling networks; adapted from CellChat https://github.com/sqjin/CellChat
}
