% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_analysis.R
\name{computeNetSimilarityPairwise_Neuron}
\alias{computeNetSimilarityPairwise_Neuron}
\title{Compute signaling network similarity for any pair of datasets}
\usage{
computeNetSimilarityPairwise_Neuron(
  object,
  slot.name = "net",
  type = c("functional"),
  comparison = NULL,
  k = NULL,
  thresh = NULL
)
}
\arguments{
\item{object}{A merged NeuronChat object}

\item{slot.name}{the slot name of object that is used to store communication strength matrices, i.e.,  'net'}

\item{type}{"functional","structural"}

\item{comparison}{a numerical vector giving the datasets for comparison}

\item{k}{the number of nearest neighbors}

\item{thresh}{the fraction (0 to 0.25) of interactions to be trimmed before computing network similarity}
}
\value{

}
\description{
Compute signaling network similarity for any pair of datasets; adapted from CellChat https://github.com/sqjin/CellChat
}
