% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NeuronChat_source_file.R
\name{createNeuronChat}
\alias{createNeuronChat}
\title{Create a new NeuronChat object from a data matrix
Create a new NeuronChat object from a data matrix; adapted from CellChat https://github.com/sqjin/CellChat}
\usage{
createNeuronChat(
  object,
  DB = c("mouse", "human"),
  meta = NULL,
  group.by = NULL,
  assay = NULL,
  do.sparse = T
)
}
\arguments{
\item{object}{a normalized (NOT count) data matrix (genes by cells)}

\item{meta}{a data frame (rows are cells with rownames) consisting of cell information, which will be used for defining cell groups.}

\item{group.by}{a vector to indicate group annotations of cells}

\item{assay}{Assay to use when the input is a Seurat object.}

\item{do.sparse}{whether use sparse format}
}
\value{

}
\description{
Create a new NeuronChat object from a data matrix
Create a new NeuronChat object from a data matrix; adapted from CellChat https://github.com/sqjin/CellChat
}
