% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{heatmap_aggregated}
\alias{heatmap_aggregated}
\title{Heatmap plot for the aggregated communication}
\usage{
heatmap_aggregated(
  object,
  method = c("weight", "count", "weighted_count", "weighted_count2",
    "weight_threshold"),
  cut_off = 0.05,
  interaction_use = "all",
  group = NULL,
  sender.names = NULL,
  receiver.names = NULL
)
}
\arguments{
\item{object}{a NeuronChat object}

\item{method}{method used for aggregation; see also function \code{net_aggregation}}

\item{cut_off}{threshold used for aggregation; see also function \code{net_aggregation}}

\item{interaction_use}{ligand-target interaction indexes  used for aggregation ('all' means use all interation pairs)}

\item{group}{a vector indicating which cell types (cell subclass, e.g., L2/3 IT) belong to which big groups (cell class, e.g., Glutamatergic)}

\item{sender.names}{sending cell groups used for plot}

\item{receiver.names}{receiving cell groups used for plot}
}
\value{

}
\description{
Heatmap plot for the aggregated communication
}
