% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{heatmap_single}
\alias{heatmap_single}
\title{Heatmap plot with ligand abundance and target abundance for single ligand-target pair}
\usage{
heatmap_single(
  object,
  interaction_name,
  group = NULL,
  sender.names = NULL,
  receiver.names = NULL
)
}
\arguments{
\item{object}{a NeuronChat object}

\item{interaction_name}{the ligand-target interaction used to plot}

\item{group}{a vector indicating which cell types (cell subclass, e.g., L2/3 IT) belong to which big groups (cell class, e.g., Glutamatergic)}

\item{sender.names}{sending cell groups used for plot}

\item{receiver.names}{receiving cell groups used for plot}
}
\value{

}
\description{
Heatmap plot with ligand abundance and target abundance for single ligand-target pair
}
