% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{lig_tar_heatmap}
\alias{lig_tar_heatmap}
\title{A set of plots illustrating the communication network as well as violin plots for related genes for single ligand-target pair
A set of plots illustrating the communication network as well as violin plots for related genes for single ligand-target pair}
\usage{
lig_tar_heatmap(
  object,
  interaction_name,
  sender.names = NULL,
  receiver.names = NULL
)
}
\arguments{
\item{object}{a NeuronChat object}

\item{interaction_name}{the ligand-target interaction used to plot}

\item{sender.names}{sending cell groups used for plot}

\item{receiver.names}{receiving cell groups used for plot}
}
\value{

}
\description{
A set of plots illustrating the communication network as well as violin plots for related genes for single ligand-target pair
A set of plots illustrating the communication network as well as violin plots for related genes for single ligand-target pair
}
