% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NeuronChat_source_file.R
\name{mergeNeuronChat}
\alias{mergeNeuronChat}
\title{Merge NeuronChat objects
Merge NeuronChat objects; adapted from CellChat https://github.com/sqjin/CellChat}
\usage{
mergeNeuronChat(
  object.list,
  add.names = NULL,
  merge.data = FALSE,
  cell.prefix = FALSE
)
}
\arguments{
\item{object.list}{A list of multiple NeuronChat objects}

\item{add.names}{A vector containing the name of each dataset}

\item{merge.data}{whether merging the data for ALL genes. Default only merges the data of signaling genes}

\item{cell.prefix}{whether prefix cell names}
}
\value{

}
\description{
Merge NeuronChat objects
Merge NeuronChat objects; adapted from CellChat https://github.com/sqjin/CellChat
}
