% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NeuronChat_source_file.R
\name{net_aggregation}
\alias{net_aggregation}
\title{Aggregation of communication networks over all ligand-target pairs}
\usage{
net_aggregation(
  net_list,
  method = c("weight", "count", "weighted_count", "weighted_count2",
    "weight_threshold"),
  cut_off = 0.05
)
}
\arguments{
\item{net_list}{a list containing communication strength matrices for all ligand-target pairs, e.g., 'net' slot of NeuronChat object after run 'run_NeuronChat'}

\item{method}{one of 'weight','count','weighted_count','weighted_count2' and 'weight_threshold'}

\item{cut_off}{threshold used for methods 'weighted_count2' and 'weight_threshold'}
}
\value{

}
\description{
Aggregation of communication networks over all ligand-target pairs
}
