% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{netVisual_chord_neuron}
\alias{netVisual_chord_neuron}
\title{Chord diagram of cell-cell communication network for single or multiple interaction pairs}
\usage{
netVisual_chord_neuron(
  object,
  method = c("weight", "count", "weighted_count", "weighted_count2",
    "weight_threshold"),
  cut_off = 0.05,
  interaction_use = "all",
  group = NULL,
  sender.names = NULL,
  receiver.names = NULL,
  lab.cex = 1.3
)
}
\arguments{
\item{object}{a NeuronChat object}

\item{method}{method used for aggregation; see also function \code{net_aggregation}}

\item{cut_off}{threshold used for aggregation; see also function \code{net_aggregation}}

\item{interaction_use}{ligand-target interaction indexes  used for aggregation ('all' means use all interation pairs); for single interaction pair, set interaction_use as the interaction name or index, e.g., 'Glu_Gria2', or simply use CellChat function netVisual_chord_cell_internal(object@net[\link{'Glu_Gria2'}], group = group,lab.cex=1.3)}

\item{group}{a vector indicating which cell types (cell subclass, e.g., L2/3 IT) belong to which big groups (cell class, e.g., Glutamatergic)}

\item{sender.names}{sending cell groups used for plot}

\item{receiver.names}{receiving cell groups used for plot}
}
\value{

}
\description{
Chord diagram of cell-cell communication network for single or multiple interaction pairs
}
