% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_analysis.R
\name{netVisual_embeddingZoomIn_Neuron}
\alias{netVisual_embeddingZoomIn_Neuron}
\title{Zoom into the 2D visualization of the learned manifold learning of the signaling networks}
\usage{
netVisual_embeddingZoomIn_Neuron(
  object,
  slot.name = "net_analysis",
  type = c("functional", "structural"),
  color.use = NULL,
  pathway.remove = NULL,
  nCol = 1,
  dot.size = c(2, 6),
  label.size = 2.8,
  dot.alpha = 0.5,
  xlabel = NULL,
  ylabel = NULL,
  do.label = T,
  show.legend = F,
  show.axes = T
)
}
\arguments{
\item{object}{NeuronChat object}

\item{slot.name}{the slot name of object that is used to store network analysis results, i.e.,  'net_analysis'}

\item{type}{"functional"}

\item{color.use}{defining the color for each cell group}

\item{pathway.remove}{a character vector defining the signaling to remove}

\item{nCol}{the number of columns of the plot}

\item{dot.size}{a range defining the size of the symbol}

\item{label.size}{font size of the text}

\item{dot.alpha}{transparency}

\item{xlabel}{label of x-axis}

\item{ylabel}{label of y-axis}

\item{do.label}{label the each point}

\item{show.legend}{whether show the legend}

\item{show.axes}{whether show the axes}
}
\value{

}
\description{
Zoom into the 2D visualization of the learned manifold learning of the signaling networks; adapted from CellChat https://github.com/sqjin/CellChat
}
