% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NeuronChat_source_file.R
\name{neuron_chat_downstream}
\alias{neuron_chat_downstream}
\title{Calculate the communication strength matrix for a single ligand-target pair (with permutation test)}
\usage{
neuron_chat_downstream(
  df,
  sender,
  receiver,
  interaction_name,
  lig_contributor,
  lig_contributor_group,
  lig_contributor_coeff,
  receptor_subunit,
  receptor_subunit_group,
  receptor_subunit_coeff,
  targets_up,
  targets_down,
  targets_nichenet,
  N,
  M,
  fdr,
  K
)
}
\arguments{
\item{df}{a data frame or data.table containing signaling gene expression (row: cell; column: genes; the last column is cell group labels)}

\item{sender}{sending cell groups}

\item{receiver}{receiving cell groups}

\item{interaction_name}{name of the ligand-target interaction}

\item{lig_contributor}{a vector with gene symbols that are regarded as ligand contributors}

\item{lig_contributor_group}{a vector indicating the grouping of ligand contributors}

\item{lig_contributor_coeff}{a vector with stoichiometry of each group of ligand contributors for calculating ligand abundance}

\item{receptor_subunit_group}{a vector indicating the grouping of target subunits}

\item{receptor_subunit_coeff}{a vector with stoichiometry of each group of target subunits for calculating target abundance}

\item{targets_up}{deprecated}

\item{targets_down}{deprecated}

\item{targets_nichenet}{deprecated}

\item{N}{deprecated}

\item{M}{number of permutation tests}

\item{fdr}{cutoff value that determines the significance of detected link}

\item{K}{deprecated}

\item{receptor_subunit_new}{a vector with updated gene symbols (removing those excluded from the expression data) that are regarded as target subunits}
}
\value{

}
\description{
Calculate the communication strength matrix for a single ligand-target pair (with permutation test)
}
