% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NeuronChat_source_file.R
\docType{class}
\name{NeuronChat-class}
\alias{NeuronChat-class}
\alias{NeuronChat}
\title{The key slots used in the NeuronChat object are described below; adapted from CellChat https://github.com/sqjin/CellChat}
\description{
The key slots used in the NeuronChat object are described below; adapted from CellChat https://github.com/sqjin/CellChat
}
\section{Slots}{

\describe{
\item{\code{data.raw}}{raw count data matrix}

\item{\code{data}}{normalized data matrix for CellChat analysis (Genes should be in rows and cells in columns)}

\item{\code{data.signaling}}{a data.frame only containing signaling genes as well as cell group labels}

\item{\code{net0}}{a list contianing the original communication strength matrix without filtered by p-value}

\item{\code{pvalue}}{a list containing the p-value (calculated by permutation test) matrices corresponding to net0}

\item{\code{net}}{a list of length K contianing the communication strength matrices (N1 X N2) filtered by p-value, where N1 is the number of sending cell groups, N2 is the number of receiving cell groups, and K is the number of ligand-target pairs. Each row of the communication strength matrix indicates the communication probability originating from the sender cell group to other cell groups}

\item{\code{net_analysis}}{a list containing various network analysis results}

\item{\code{DB}}{ligand-target interaction database used in the analysis}

\item{\code{LR}}{a list of information related with ligand-target pairs}

\item{\code{meta}}{data frame storing the information associated with each cell}

\item{\code{idents}}{a factor defining the cell identity used for all analysis. It becomes a list for a merged NeuronChat object}

\item{\code{dr}}{List of the reduced 2D coordinates, one per method, e.g., umap/tsne/dm}

\item{\code{options}}{List of miscellaneous data, such as parameters used throughout analysis, and a indicator whether the NeuronChat object is a single or merged}

\item{\code{fc}}{deprecated}

\item{\code{info}}{information flow of each ligand-target pair}

\item{\code{ligand.abundance}}{ligand.abundance of sending cell groups (row) for ligand-target pairs (column)}

\item{\code{target.abundance}}{target.abundance of receiving cell groups (row) for ligand-target pairs (column)}
}}

