% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_analysis.R
\name{selectK_Neuron}
\alias{selectK_Neuron}
\title{Select the number of the patterns for running \code{identifyCommunicationPatterns}}
\usage{
selectK_Neuron(
  object,
  slot.name = "net",
  pattern = c("outgoing", "incoming"),
  title.name = NULL,
  do.facet = TRUE,
  k.range = seq(2, 10),
  nrun = 30,
  seed.use = 10
)
}
\arguments{
\item{object}{NeuronChat object}

\item{pattern}{"outgoing" or "incoming"}

\item{title.name}{title of plot}

\item{do.facet}{whether use facet plot showing the two measures}

\item{k.range}{a range of the number of patterns}

\item{nrun}{number of runs when performing NMF}

\item{seed.use}{seed when performing NMF}

\item{slot.namethe}{slot name of object that is used to store communication strength matrices, i.e.,  'net'}
}
\value{
a ggplot object
}
\description{
Select the number of the patterns for running \code{identifyCommunicationPatterns}; adapted from CellChat https://github.com/sqjin/CellChat
We infer the number of patterns based on two metrics that have been implemented in the NMF R package, including Cophenetic and Silhouette. Both metrics measure the stability for a particular number of patterns based on a hierarchical clustering of the consensus matrix. For a range of the number of patterns, a suitable number of patterns is the one at which Cophenetic and Silhouette values begin to drop suddenly.
}
