% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_analysis_visualization.R
\name{spatialDistribution}
\alias{spatialDistribution}
\title{Spatial distribution of cell types along sptial axis}
\usage{
spatialDistribution(
  meta,
  celltype_label = "subclass",
  centroid = "centroid_y",
  curve.alpha = 0.5,
  font.size = 20,
  legend.symbol.size = 5
)
}
\arguments{
\item{meta}{A dataframe containing spatial locations and cell type annotations of cells}

\item{celltype_label}{variable name (character) in meta, to indicate the cell type annotations of cells}

\item{centroid}{variable name (character) in meta, to indicate the  x or y coordinates of cells}
}
\value{
a ggplot object
}
\description{
Spatial distribution of cell types along sptial axis
}
