% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_analysis_visualization.R
\name{spatialMap}
\alias{spatialMap}
\title{Spatial plot of cells with their annotations}
\usage{
spatialMap(
  meta,
  celltype_label = "subclass",
  centroid_x = "centroid_x",
  centroid_y = "centroid_y",
  pt.size = 5,
  pt.alpha = 0.8,
  font.size = 20,
  legend.symbol.size = 5
)
}
\arguments{
\item{meta}{A dataframe containing spatial locations and cell type annotations of cells}

\item{celltype_label}{variable name (character) in meta, to indicate the cell type annotations of cells}

\item{centroid_x}{variable name (character) in meta, to indicate the x coordinates of cells}

\item{centroid_y}{variable name (character) in meta, to indicate the y coordinates of cells}
}
\value{
a ggplot object
}
\description{
Spatial plot of cells with their annotations
}
