% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_NeuronChatDB.R
\name{update_interactionDB}
\alias{update_interactionDB}
\title{Update NeuronChatDB using user defined new entry information}
\usage{
update_interactionDB(
  DB,
  interaction_name,
  lig_contributor,
  receptor_subunit,
  interaction_type = "user_defined",
  ligand_type = "user_defined",
  lig_contributor_group = NULL,
  lig_contributor_coeff = NULL,
  receptor_subunit_group = NULL,
  receptor_subunit_coeff = NULL
)
}
\arguments{
\item{DB}{current interation database (stored in slot 'DB' of NeuronChat object, i.e., object@DB)}

\item{interaction_name}{a character to denote the name of the interaction to be added}

\item{lig_contributor}{a vector of gene symbols, contributing to ligand abundance}

\item{receptor_subunit}{a vector of gene symbols, contributing to receptor abundance}

\item{interaction_type}{a character to denote the interaction type}

\item{ligand_type}{a character to denote the ligand_type}

\item{lig_contributor_group}{a vector of integers to indicate the biological function group of genes in \code{lig_contributor};
the length of \code{lig_contributor_group} is the same as \code{lig_contributor}}

\item{lig_contributor_coeff}{a numerical vector to indicate the stoichiometric coefficients of ligand-contribuing groups, to calculate ligand abundance;
the length is the same as the unique of \code{lig_contributor_group}}

\item{receptor_subunit_group}{a vector of integers to indicate the biological function group of genes in \code{receptor_subunit};
the length of \code{receptor_subunit_group} is the same as \code{receptor_subunit}}

\item{receptor_subunit_coeff}{a numerical vector to indicate the stoichiometric coefficients of receptor subunit groups, to calculate receptor abundance;
the length is the same as the unique of \code{receptor_subunit_group}}
}
\value{
a updated interaction database (a list)
}
\description{
Update NeuronChatDB using user defined new entry information
}
