% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NeuronChat_source_file.R
\name{run_NeuronChat}
\alias{run_NeuronChat}
\title{Calculate the communication strength matrix for all ligand-target pairs}
\usage{
run_NeuronChat(
  object = object,
  sender = NULL,
  receiver = NULL,
  N = 0,
  M = 100,
  fdr = 0.05,
  K = 0.1
)
}
\arguments{
\item{object}{a NeuronChat object}

\item{sender}{sending cell groups}

\item{receiver}{receiving cell groups}

\item{N}{deprecated}

\item{M}{number of permutation tests}

\item{fdr}{cutoff value that determines the significance of detected link}

\item{K}{deprecated}
}
\value{

}
\description{
Calculate the communication strength matrix for all ligand-target pairs
}
